functions/Naive_Bayes_Agrawal/test/requirements.txt
pulsar-client==2.4.1.post1
scikit-multiflow==0.5.3
numpy==1.18.5